Proteção das rotas de admin utilizando middleware

+++ routes/web.php
@@ -31,7 +31,7 @@ Route::post('/login',[LoginController::class, 'authenticate'])->name('login');
 Route::post('/logout',[LoginController::class, 'logout'])->name('logout');
 
 /* Links da sidebar */
-Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard.index')->middleware('auth');
+Route::get('/dashboard', [DashboardController::class, 'index'])->name('dashboard.index')->middleware('auth.admin');
 Route::get('/carimbos/b2b', function(){ return view('fallback.manutencao');})->name('carimbos.b2b.index')->middleware('auth');
 Route::get('/carimbos/controle', [ControleController::class, 'carregarViewControle'])->name('carimbos.controle.index')->middleware('auth');
 Route::get('/carimbos/gerais', function(){ return view('fallback.manutencao');})->name('carimbos.gerais.index')->middleware('auth');
@@ -49,15 +49,15 @@ Route::post('/carimbos/controle/formularios/controle_crise/insert',[ControleCont
 /* Fim do insert de carimbos */
 
 /* CRUD de usuários */
-Route::get('/usuarios', [UsuarioController::class, 'index'])->name('usuarios.index')->middleware('auth');
-Route::post('/usuarios/lista_usuarios', [UsuarioController::class, 'listarUsuarios'])->name('usuarios.listar')->middleware('auth');
-Route::get('/usuarios/create', [UsuarioController::class, 'create'])->name('usuarios.create')->middleware('auth');
-Route::post('/usuarios/store', [UsuarioController::class, 'store'])->name('usuarios.store')->middleware('auth');
-Route::get('/usuarios/{user}', [UsuarioController::class, 'show'])->name('usuarios.show')->middleware('auth');
-Route::post('/usuarios', [UsuarioController::class, 'listarAtividadesUsuario'])->name('usuarios.listar_atividades')->middleware('auth');
-Route::get('/usuarios/{user}/edit', [UsuarioController::class, 'edit'])->name('usuarios.edit')->middleware('auth');
-Route::put('/usuarios/{user}/update', [UsuarioController::class, 'update'])->name('usuarios.update')->middleware('auth');
-Route::delete('/usuarios/{user}/destroy', [UsuarioController::class, 'destroy'])->name('usuarios.delete')->middleware('auth');
+Route::get('/usuarios', [UsuarioController::class, 'index'])->name('usuarios.index')->middleware('auth.admin');
+Route::post('/usuarios/lista_usuarios', [UsuarioController::class, 'listarUsuarios'])->name('usuarios.listar')->middleware('auth.admin');
+Route::get('/usuarios/create', [UsuarioController::class, 'create'])->name('usuarios.create')->middleware('auth.admin');
+Route::post('/usuarios/store', [UsuarioController::class, 'store'])->name('usuarios.store')->middleware('auth.admin');
+Route::get('/usuarios/{user}', [UsuarioController::class, 'show'])->name('usuarios.show')->middleware('auth.admin');
+Route::post('/usuarios', [UsuarioController::class, 'listarAtividadesUsuario'])->name('usuarios.listar_atividades')->middleware('auth.admin');
+Route::get('/usuarios/{user}/edit', [UsuarioController::class, 'edit'])->name('usuarios.edit')->middleware('auth.admin');
+Route::put('/usuarios/{user}/update', [UsuarioController::class, 'update'])->name('usuarios.update')->middleware('auth.admin');
+Route::delete('/usuarios/{user}/destroy', [UsuarioController::class, 'destroy'])->name('usuarios.delete')->middleware('auth.admin');
 /* Fim do CRUD de usuários */
 
 //Fallback(tratar requisições a rotas inexistentes)
